+++ src/main/java/com/github/siboxd/fatturapa/model/invoiceheader/supplier/ContattiType.java
@@ -1,22 +1,27 @@
-package com.github.siboxd.fatturapa.model;
+package com.github.siboxd.fatturapa.model.invoiceheader.supplier;
 
 import org.simpleframework.xml.Element;
 import org.simpleframework.xml.Root;
 
 
 /**
- * ContattiType<br>
+ * Contains contact information for a subject.
+ * <p>
  * Generated using Android JAXB<br>
  *
  * @link https://github.com/yeshodhan/android-jaxb
  */
 @Root(name = "ContattiType")
 public class ContattiType {
+    // TODO: 19/01/2019 is that class the same as ContattiTrasmittenteType???
+    // ...maybe it could be an extension...
 
     @Element(name = "Telefono", required = false)
     private String telefono;
+
     @Element(name = "Fax", required = false)
     private String fax;
+
     @Element(name = "Email", required = false)
     private String email;
 
@@ -27,7 +32,10 @@ public class ContattiType {
         return telefono;
     }
 
-    public void setTelefono(String telefono) {
+    /**
+     * @param telefono The field, if evaluated, must contain a telephone number of the subject.
+     */
+    public void setTelefono(final String telefono) {
         this.telefono = telefono;
     }
 
@@ -35,7 +43,10 @@ public class ContattiType {
         return fax;
     }
 
-    public void setFax(String fax) {
+    /**
+     * @param fax The field, if evaluated, must contain a fax number of the subject.
+     */
+    public void setFax(final String fax) {
         this.fax = fax;
     }
 
@@ -43,7 +54,10 @@ public class ContattiType {
         return email;
     }
 
-    public void setEmail(String email) {
+    /**
+     * @param email The field, if evaluated, must contain an e-mail address of the subject.
+     */
+    public void setEmail(final String email) {
         this.email = email;
     }
